숫자 버튼과 버튼 그룹 분리 완료
