+++ src/component/number-button.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function NumberButton({ count, number, onClick }) {
+  return (
+    <button type="button" onClick={() => onClick({ count, number })}>
+      {number}
+    </button>
+  );
+}
+
+export default NumberButton;
