+++ src/index.jsx
@@ -2,29 +2,7 @@ import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 import Title from './component/title';
 import MainButton from './component/main-button';
-
-function ButtonGroup({ count, onClick }) {
-  const numbers = [1, 2, 3, 4, 5];
-  return (
-    <p>
-      {numbers.map((number) => (
-        <NumberButton
-          count={count}
-          number={number}
-          onClick={onClick}
-        />
-      ))}
-    </p>
-  );
-}
-
-function NumberButton({ count, number, onClick }) {
-  return (
-    <button type="button" onClick={() => onClick({ count, number })}>
-      {number}
-    </button>
-  );
-}
+import ButtonGroup from './component/button-group';
 
 function App() {
   const [state, setState] = useState({
